+++ internal/sandbox/handler.go
@@ -150,7 +150,7 @@ func (h *SetStateHandler) parseCtx(ctx context.Context, manager runtime.RuntimeM
 	var errs = &types.SandboxDataErrors{}
 
 	// FIXME do full parsing once independent parsing is implemented
-	_, err := policy.Unmarshal(data.PolicyDefinition, types.PolicyMarshalingType_SHORT_YAML)
+	_, err := policy.Unmarshal(data.PolicyDefinition, types.PolicyMarshalingType_YAML)
 	if err != nil {
 		end := getPolicyEndPosition(data.PolicyDefinition)
 		err := &types.LocatedMessage{
@@ -219,7 +219,7 @@ func (h *SetStateHandler) setPolicy(ctx context.Context, manager runtime.Runtime
 	polHandler := policy.CreatePolicyHandler{}
 	polResp, err := polHandler.Execute(authenticatedCtx, manager, &types.CreatePolicyRequest{
 		Policy:      simCtx.PolicyDefinition,
-		MarshalType: types.PolicyMarshalingType_SHORT_YAML,
+		MarshalType: types.PolicyMarshalingType_YAML,
 	})
 
 	if err != nil {