+++ test/benchmark/engine/relationship/create_test.go
@@ -12,17 +12,17 @@ import (
 	"github.com/sourcenetwork/acp_core/test"
 )
 
-var policyStr string = `
-name: policy
+var policyStr string = `name: policy
 resources:
-  doc:
-    relations:
-      owner:
-        types:
-          - actor
-      reader:
-        types:
-          - actor
+- name: doc
+  relations:
+  - name: owner
+    types:
+    - actor
+  - name: reader
+    types:
+    - actor
+spec: none
 `
 
 func BenchmarkCreateRelationshipsWithObjects(b *testing.B) {