+++ test/integration/playground/simulate_test.go
@@ -13,19 +13,19 @@ func Test_Simulate_ReturnsErrorIfSubjectInOwnerRelationshipIsNotActor(t *testing
 	ctx := test.NewTestCtx(t)
 
 	data := types.SandboxData{
-		PolicyDefinition: `
-name: shinzo
+		PolicyDefinition: `name: shinzo
 resources:
-  file:
-    relations:
-      owner:
-		types:
-        - actor
-  group:	
-    relations:
-      owner:
-        types:
-        - actor
+- name: file
+  relations:
+  - name: owner
+    types:
+    - actor
+- name: group
+  relations:
+  - name: owner
+    types:
+    - actor
+spec: none
 `,
 		Relationships: `
 file:logs#owner@group:example#owner